requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.3.0=mkl
abseil-cpp=20211102.0=hd77b12b_0
absl-py=0.15.0=pyhd3eb1b0_0
aiohttp=3.8.1=py38h2bbff1b_1
aiosignal=1.2.0=pyhd3eb1b0_0
altair=4.1.0=py_1
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py38h2bbff1b_0
arrow-cpp=8.0.0=py38hbd6f097_0
asttokens=2.0.5=pyhd3eb1b0_0
astunparse=1.6.3=py_0
async-timeout=4.0.2=py38haa95532_0
attrs=21.4.0=pyhd3eb1b0_0
aws-c-common=0.4.57=ha925a31_1
aws-c-event-stream=0.1.6=hd77b12b_5
aws-checksums=0.1.9=ha925a31_0
aws-sdk-cpp=1.8.185=hd77b12b_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py38haa95532_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.4=py38haa95532_0
boost-cpp=1.73.0=h2bbff1b_12
bottleneck=1.3.5=py38h080aedc_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotlipy=0.7.0=py38h2bbff1b_1003
bzip2=1.0.8=he774522_0
c-ares=1.18.1=h2bbff1b_0
ca-certificates=2022.07.19=haa95532_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2022.9.14=py38haa95532_0
cffi=1.15.1=py38h2bbff1b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py38haa95532_0
colorama=0.4.5=py38haa95532_0
commonmark=0.9.1=pyhd3eb1b0_0
cryptography=37.0.1=py38h21b164f_0
cycler=0.11.0=pyhd3eb1b0_0
dataclasses=0.8=pyh6d0b6a4_7
debugpy=1.5.1=py38hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
eigen=3.3.7=h59b6b97_1
entrypoints=0.4=py38haa95532_0
executing=0.8.3=pyhd3eb1b0_0
fftw=3.3.9=h2bbff1b_1
flatbuffers=2.0.0=h6c2663c_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
frozenlist=1.2.0=py38h2bbff1b_0
future=0.18.2=py38_1
gast=0.5.3=pyhd3eb1b0_0
gflags=2.2.2=ha925a31_0
giflib=5.2.1=h62dcd97_0
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.18=pyhd3eb1b0_1
glib=2.69.1=h5dc1a3c_1
glog=0.5.0=hd77b12b_0
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-pasta=0.2.0=pyhd3eb1b0_0
grpcio=1.42.0=py38hc60d5dd_0
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
h5py=3.7.0=py38h3de5c98_0
hdf5=1.10.6=h1756f20_1
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=py38haa95532_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.15.2=py38haa95532_0
ipython=8.4.0=py38haa95532_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py38haa95532_1
jinja2=3.0.3=pyhd3eb1b0_0
jpeg=9e=h2bbff1b_0
jsonschema=4.16.0=py38haa95532_0
jupyter_client=7.3.5=py38haa95532_0
jupyter_core=4.10.0=py38haa95532_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
keras=2.8.0=py38haa95532_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.4.2=py38hd77b12b_0
lerc=3.0=hd77b12b_0
libboost=1.73.0=h6c2663c_12
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=12.0.0=default_h627e005_2
libcurl=7.84.0=h86230a5_0
libdeflate=1.8=h2bbff1b_5
libffi=3.4.2=hd77b12b_4
libiconv=1.16=h2bbff1b_2
libogg=1.3.5=h2bbff1b_1
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.20.1=h23ce68f_0
libsodium=1.0.18=h62dcd97_0
libssh2=1.10.0=hcd4344a_0
libthrift=0.15.0=he1d8c1a_0
libtiff=4.4.0=h8a3f274_0
libvorbis=1.3.7=he774522_0
libwebp=1.2.2=h2bbff1b_0
libwebp-base=1.2.2=h2bbff1b_0
libxml2=2.9.14=h0ad7f3c_0
libxslt=1.1.35=h2bbff1b_0
lz4-c=1.9.3=h2bbff1b_1
markdown=3.3.4=py38haa95532_0
markupsafe=2.1.1=py38h2bbff1b_0
matplotlib=3.5.2=py38haa95532_0
matplotlib-base=3.5.2=py38hd77b12b_0
matplotlib-inline=0.1.6=py38haa95532_0
mistune=0.8.4=py38he774522_1000
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.1=py38h277e83a_0
mkl_random=1.2.2=py38hf11a4ad_0
multidict=5.2.0=py38h2bbff1b_3
munkres=1.1.4=py_0
nbclient=0.5.13=py38haa95532_0
nbconvert=6.4.4=py38haa95532_0
nbformat=5.5.0=py38haa95532_0
nest-asyncio=1.5.5=py38haa95532_0
notebook=6.4.12=py38haa95532_0
numexpr=2.8.3=py38hb80d3ca_0
numpy=1.23.1=py38h7a0a035_0
numpy-base=1.23.1=py38hca35cd5_0
oauthlib=3.2.0=pyhd3eb1b0_1
opencv=4.6.0=py38h104de81_2
openssl=1.1.1q=h2bbff1b_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.4=py38hd77b12b_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=hd77b12b_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py38hdc2b20a_1
pip=22.2.2=py38haa95532_0
pkgutil-resolve-name=1.3.10=py38haa95532_0
ply=3.11=py38_0
prometheus_client=0.14.1=py38haa95532_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.20.1=py38hd77b12b_0
psutil=5.9.0=py38h2bbff1b_0
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=8.0.0=py38h953b917_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pydeck=0.7.1=py38haa95532_0
pygments=2.11.2=pyhd3eb1b0_0
pyjwt=2.4.0=py38haa95532_0
pympler=0.9=py_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py38haa95532_0
pyqt=5.15.7=py38hd77b12b_0
pyqt5-sip=12.11.0=py38hd77b12b_0
pyrsistent=0.18.0=py38h196d8e1_0
pysocks=1.7.1=py38haa95532_0
python=3.8.13=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py38haa95532_0
python-flatbuffers=2.0=pyhd3eb1b0_0
pytz=2022.1=py38haa95532_0
pywin32=302=py38h2bbff1b_2
pywinpty=2.0.2=py38h5da7b33_0
pyyaml=6.0=py38h2bbff1b_1
pyzmq=23.2.0=py38hd77b12b_0
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=hb9a9bb5_4
qtwebkit=5.212=h3ad3cdb_4
re2=2022.04.01=hd77b12b_0
requests=2.28.1=py38haa95532_0
requests-oauthlib=1.3.0=py_0
rich=12.5.1=py38haa95532_0
rsa=4.7.2=pyhd3eb1b0_1
scipy=1.9.1=py38he11b74f_0
semver=2.13.0=pyhd3eb1b0_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=63.4.1=py38haa95532_0
simplemma=0.5.0=pypi_0
sip=6.6.2=py38hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
smmap=4.0.0=pyhd3eb1b0_0
snappy=1.1.9=h6c2663c_0
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.39.3=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
streamlit=1.11.0=py38haa95532_0
tensorboard=2.8.0=py38haa95532_0
tensorboard-data-server=0.6.0=py38haa95532_0
tensorboard-plugin-wit=1.8.1=py38haa95532_0
tensorflow=2.8.2=mkl_py38h6f30489_0
tensorflow-base=2.8.2=mkl_py38h6a7f48e_0
tensorflow-estimator=2.8.0=py38hd4e2768_0
tensorflow-hub=0.12.0=pypi_0
termcolor=1.1.0=py38haa95532_1
terminado=0.13.1=py38haa95532_0
testpath=0.6.0=py38haa95532_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.11.2=pyhd3eb1b0_0
tornado=6.2=py38h2bbff1b_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.3.0=py38haa95532_0
typing_extensions=4.3.0=py38haa95532_0
tzlocal=2.1=py38_0
urllib3=1.26.11=py38haa95532_0
utf8proc=2.6.1=h2bbff1b_0
validators=0.18.2=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=2.1.6=py38haa95532_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py38haa95532_0
win_inet_pton=1.1.0=py38haa95532_0
wincertstore=0.2=py38haa95532_2
winpty=0.4.3=4
wrapt=1.14.1=py38h2bbff1b_0
xz=5.2.6=h8cc25b3_0
yaml=0.2.5=he774522_0
yarl=1.8.1=py38h2bbff1b_0
zeromq=4.3.4=hd77b12b_0
zipp=3.8.0=py38haa95532_0
zlib=1.2.12=h8cc25b3_3
zstd=1.5.2=h19a0ad4_0
